Add separators between character colour pickers
We wanted to make it clearer which colour the text related to on the
character colour picker. I tried a few different things to increase the
spacing (adding white space, trying to target it with CSS) but the only
thing that had an effect is adding pipe separators.

Suspect this may just be a limitation with Blockly.